Update Sample_Script.ps1
Added a series of commands which I've been using to test the module. I've tested them with the web service and the COM. It will create a new VM named OtherTest64 and two virtual disks, make a series of modifications to them, ending with everything created being deleted from the host. **IF YOU HAVE A VM WITH A NAME MATCHING "othertest" OR VIRTUAL DISKS WITH NAMES MATCHING "testdisk" THEY WILL BE DELETED!**